@@ -314,17 +314,17 @@
     "svg-react-loader": "^0.4.5",
     "tape": "^4.4.0",
     "thread-loader": "^1.1.2",
-    "ts-jest": "^22.4.4",
-    "ts-loader": "^4.2.0",
-    "ts-node": "^6.0.0",
-    "tsconfig-paths": "^3.3.1",
-    "tslint": "^5.9.1",
+    "ts-jest": "^22.4.6",
+    "ts-loader": "^4.3.0",
+    "ts-node": "^6.0.3",
+    "tsconfig-paths": "^3.3.2",
+    "tslint": "^5.10.0",
     "typescript": "^2.8.3",
     "typescript-formatter": "^7.2.0",
     "url-loader": "^1.0.1",
-    "webpack": "^4.6.0",
-    "webpack-cli": "^2.0.15",
-    "webpack-dev-server": "^3.1.3",
+    "webpack": "^4.8.3",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-server": "^3.1.4",
     "yarn-run-all": "^3.1.1"
   }
 }
