@@ -48,8 +48,8 @@ export class JobConfig
 {
   public createdAt: Date = null;                   // when the job was created
   public createdBy: number = null;                 // who created the job (user ID)
-  public id?: number;                              // job ID
-  public logId?: number;                           // jobLog table ID
+  public id?: number = null;                       // job ID
+  public logId?: number = null;                    // jobLog table ID
   public meta: string = '';                        // meta
   public name: string = '';                        // name of the job
   public pausedFilename: string = '';              // filename of the stored paused output
