@@ -506,7 +506,9 @@ class PathfinderFilterSection extends TerrainComponent<Props>
           />
           {
             filterGroup.lines.map((line, i) =>
-              <div key={i}>
+              <div
+                key={line.id === -1 ? i : line.id /* dealing with id-less lines */}
+              >
                 {
                   !this.isGroup(line) ?
                     <DragDropItem
