@@ -43,91 +43,18 @@ THE SOFTWARE.
 */
 
 // Copyright 2018 Terrain Data, Inc.
-// tslint:disable:no-var-requires
 import TerrainComponent from 'common/components/TerrainComponent';
-import * as Immutable from 'immutable';
-import * as _ from 'lodash';
-import memoizeOne from 'memoize-one';
-import * as Radium from 'radium';
 import * as React from 'react';
+import './Badge.less';
 
-import FadeInOut from 'common/components/FadeInOut';
-
-import { instanceFnDecorator } from 'shared/util/Classes';
-import { backgroundColor, borderColor, Colors, fontColor, getStyle } from 'src/app/colors/Colors';
-import Util from 'util/Util';
-
-import TemplateList, { AllowedActions } from 'etl/templates/components/TemplateList';
-import { WalkthroughActions } from 'etl/walkthrough/ETLWalkthroughRedux';
-import { ViewState, WalkthroughState } from 'etl/walkthrough/ETLWalkthroughTypes';
-import { _SinkConfig, _SourceConfig, SinkConfig, SourceConfig } from 'shared/etl/immutable/EndpointRecords';
-import { ETLTemplate } from 'shared/etl/immutable/TemplateRecords';
-
-import { ETLStepComponent, StepProps, TransitionParams } from './ETLStepComponent';
-import './ETLStepComponent.less';
-
-const loadTemplateActions: AllowedActions = {
-  delete: true,
-};
-
-class PickTemplateStep extends ETLStepComponent
+interface BadgeProps
 {
-  public static onRevert(params: TransitionParams)
-  {
-    params.act({
-      actionType: 'setState',
-      state: {
-        chosenTemplateId: -1,
-      },
-    });
-  }
-
-  public getTemplateItemStyle(template)
-  {
-    return templateListItemStyle;
-  }
-
-  public render()
-  {
-    return (
-      <div
-        style={templateListStyle}
-      >
-        <TemplateList
-          onClick={this.handleLoadTemplateItemClicked}
-          getRowStyle={this.getTemplateItemStyle}
-          allowedActions={loadTemplateActions}
-        />
-      </div>
-    );
-  }
-
-  public handleLoadTemplateItemClicked(template: ETLTemplate)
-  {
-    const { onDone, act } = this.props;
-    act({
-      actionType: 'setState',
-      state: {
-        chosenTemplateId: template.id,
-      },
-    });
-    onDone();
-  }
+  label: string;
+  color: string;
 }
 
-const templateListItemStyle = [
-  { cursor: 'pointer' },
-  backgroundColor('rgba(0,0,0,0)', Colors().activeHover),
-];
-
-const templateListStyle = _.extend({},
-  backgroundColor(Colors().bg3),
-  getStyle('boxShadow', `2px 2px 3px ${Colors().boxShadow}`),
+const Badge = (props: BadgeProps) => (
+  <div className='badge' style={{ backgroundColor: props.color }}>{props.label}</div>
 );
 
-const transitionRowHeight = '28px';
-export default Util.createTypedContainer(
-  PickTemplateStep,
-  ['walkthrough'],
-  { act: WalkthroughActions },
-);
+export default Badge;
