@@ -43,6 +43,7 @@ THE SOFTWARE.
 */
 
 // Copyright 2018 Terrain Data, Inc.
+import { ModalProps, MultiModal } from 'common/components/overlay/MultiModal';
 import TerrainComponent from 'common/components/TerrainComponent';
 import TerrainTabs from 'common/components/TerrainTabs';
 import { ETLActions } from 'etl/ETLRedux';
@@ -64,6 +65,7 @@ interface DataTabsProps
   params: any;
   router: any;
   children: JSX.Element;
+  modalRequests: List<ModalProps>;
 }
 
 class DataTabs extends TerrainComponent<DataTabsProps>
@@ -74,7 +76,7 @@ class DataTabs extends TerrainComponent<DataTabsProps>
     { key: 'schedules', label: 'Schedules' },
     { key: 'jobs', label: 'Jobs' },
     
-    { key: 'runnow', label: 'Import / Export Now' },
+    { key: 'newtemplate', label: 'New Import or Export' },
   ];
 
   public tabToRouteMap = {
@@ -83,11 +85,7 @@ class DataTabs extends TerrainComponent<DataTabsProps>
     schedules: '/data/schedules',
     jobs: '/data/jobs',
     
-    runnow: '/data/runnow',
-  };
-
-  public state = {
-    tabIndex: this.tabs.indexOf(this.props.params.tab),
+    newtemplate: '/data/newtemplate',
   };
 
   public componentDidMount()
@@ -102,9 +100,17 @@ class DataTabs extends TerrainComponent<DataTabsProps>
     // TODO lock UI until done?
   }
 
+  public setModalRequests(requests)
+  {
+    this.props.etlActions({
+      actionType: 'setModalRequests',
+      requests,
+    });
+  }
+
   public render()
   {
-    const { params, router, children } = this.props;
+    const { params, router, children, modalRequests } = this.props;
 
     return (
       <div className='etl'>
@@ -118,6 +124,10 @@ class DataTabs extends TerrainComponent<DataTabsProps>
               this.props.children
             }
           </TerrainTabs>
+          <MultiModal
+            requests={modalRequests}
+            setRequests={this.setModalRequests}
+          />
         </div>
       </div>
     );
@@ -126,6 +136,6 @@ class DataTabs extends TerrainComponent<DataTabsProps>
 
 export default Util.createContainer(
   DataTabs,
-  [],
+  [['etl', 'modalRequests']],
   { etlActions: ETLActions },
 );
