@@ -58,9 +58,10 @@ const { List, Map } = Immutable;
 import { ArraySumTFF } from './ArraySumTransformationForm';
 import { CastTFF } from './CastTransformationForm';
 import { DuplicateTFF } from './DuplicateTransformationForm';
+import { InsertTFF } from './InsertTransformationForm';
 import { JoinTFF } from './JoinTransformationForm';
 import { SetIfTFF } from './SetIfTransformationForm';
-import { AddTFF, DivideTFF, HashTFF, MultiplyTFF, SubstringTFF, SubtractTFF, UppercaseTFF } from './SimpleTransformations';
+import { AddTFF, DivideTFF, FindReplaceTFF, HashTFF, MultiplyTFF, SubstringTFF, SubtractTFF, UppercaseTFF } from './SimpleTransformations';
 import { SplitTFF } from './SplitTransformationForm';
 
 export function getTransformationForm(type: TransformationNodeType): React.ComponentClass<TransformationFormProps>
@@ -93,6 +94,10 @@ export function getTransformationForm(type: TransformationNodeType): React.Compo
       return DivideTFF;
     case TransformationNodeType.SetIfNode:
       return SetIfTFF;
+    case TransformationNodeType.FindReplaceNode:
+      return FindReplaceTFF;
+    case TransformationNodeType.InsertNode:
+      return InsertTFF;
     default:
       return null;
   }
