@@ -47,12 +47,19 @@ import { createSelector } from 'reselect';
 
 const getJobs = (state) => state.get('jobs').jobs;
 
-export const getSuccessfulJobs = createSelector(
+export const getCompletedJobs = createSelector(
   getJobs,
-  (jobs) => jobs.filter((j) => j.status === 'SUCCESS'),
+  (jobs) => jobs.filter((j) =>
+    ['SUCCESS', 'FAILURE', 'ABORTED', 'CANCELED'].indexOf(j.status) > -1,
+  ),
 );
 
-export const getFailedJobs = createSelector(
+export const getPendingJobs = createSelector(
   getJobs,
-  (jobs) => jobs.filter((j) => j.status === 'FAILURE'),
+  (jobs) => jobs.filter((j) => j.status === 'PENDING'),
+);
+
+export const getRunningJobs = createSelector(
+  getJobs,
+  (jobs) => jobs.filter((j) => j.status === 'RUNNING'),
 );
